# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: AquaTracker API v1
  description:  >
    AquaTracker REST API. All request/response bodies are application/json.
    Base URL: http://localhost:3001/api/v1/
  version: 1.0.0
servers:
  - url: http://localhost:3001/api/v1
    description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: temperatureC must be between 15 and 35
        field:
          type: string
          nullable: true
          example: temperatureC
        details:
          type: object
          nullable: true
          additionalProperties: true
          example:
            fishId: fish_betta
            conflictingFishIds: [fish_betta_male_2]
      required: [error, message]
    Language:
      type: string
      enum: [pl, en]
    Theme:
      type: string
      enum: [light, dark]
    WaterType:
      type: string
      enum: ["Słodkowodna", "Słonawowodna", "Słonowodna"]
...
    Fish:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          example: fish_001
        name:
          type: string
          example: Neon Innesa
        waterType:
          $ref: "#/components/schemas/WaterType"
        temperature:
          type: string
          description: Temperature range string, e.g. "22-26"
          example: "22-26"
        biotope:
          type: string
          example: Ameryka Południowa
        ph:
          type: string
          example: "6.5-7.5"
        hardnessDGH:
          type: string
          example: "1-12"
        temperament:
          type: string
          example: spokojne
        minShoalSize:
          type: integer
          minimum: 1
          example: 10
        lifeSpan:
          type: string
          example: 3-5 lat
        iconName:
          type: string
          example: neon_innesa.png
      required:
        - id
        - name
        - waterType
        - temperature
        - biotope
        - ph
        - hardnessDGH
        - temperament
        - minShoalSize
        - lifeSpan
        - iconName
    FishCreateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        waterType:
          $ref: "#/components/schemas/WaterType"
        temperature:
          type: string
        biotope:
          type: string
        ph:
          type: string
        hardnessDGH:
          type: string
        temperament:
          type: string
        minShoalSize:
          type: integer
          minimum: 1
        lifeSpan:
          type: string
        iconName:
          type: string
      required:
        - name
        - waterType
        - temperature
        - biotope
        - ph
        - hardnessDGH
        - temperament
        - minShoalSize
        - lifeSpan
        - iconName
...
    Settings:
      type: object
      additionalProperties: false
      properties:
        language:
          $ref: "#/components/schemas/Language"
        theme:
          $ref: "#/components/schemas/Theme"
        sessionLengthMinutes:
          type: integer
          minimum: 1
          example: 60
        dataSource:
          type: string
          example: production
      required: [language, theme, sessionLengthMinutes, dataSource]
    SettingsUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        language:
          $ref: "#/components/schemas/Language"
        theme:
          $ref: "#/components/schemas/Theme"
        sessionLengthMinutes:
          type: integer
          minimum: 1
        dataSource:
          type: string
...
paths:
...
  /settings:
    get:
      tags: [Settings]
      summary: Get settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
              example:
                language: pl
                theme: dark
                sessionLengthMinutes: 60
                dataSource: production
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags: [Settings]
      summary: Update settings (partial)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdateRequest"
            example:
              language: en
              theme: dark
              sessionLengthMinutes: 90
              dataSource: demo
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /fish:
    get:
      tags: [Fish]
      summary: List fish (with filters)
      security: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Name substring filter.
        - name: waterType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/WaterType"
        - name: temperament
          in: query
          required: false
          schema:
            type: string
          example: spokojne
        - name: tempMin
          in: query
          required: false
          schema:
            type: number
          description: Minimum temperature (°C).
        - name: tempMax
          in: query
          required: false
          schema:
            type: number
          description: Maximum temperature (°C).
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Fish list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fish"

    post:
      tags: [Fish]
      summary: Add fish
      description: Creates a fish entry (admin/internal).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FishCreateRequest"
      responses:
        "201":
          description: Created fish
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fish"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...